patch 7.4.1863 Problem:    Compiler warnings on Win64. Solution:   Adjust types, add type casts. (Ken Takata)

+++ src/if_ruby.c
@@ -733,7 +733,7 @@ vim_str2rb_enc_str(const char *s)
 	vim_free(sval);
 	if (enc)
 	{
-	    return rb_enc_str_new(s, strlen(s), enc);
+	    return rb_enc_str_new(s, (long)strlen(s), enc);
 	}
     }
 #endif
